+++ src/main/java/com/island/engine/ecs/EntitySystem.java
@@ -5,5 +5,11 @@ import com.island.engine.model.Mortal;
 import com.island.engine.service.CellService;
 
 public interface EntitySystem<T extends Mortal> extends CellService<T> {
-    List<Class<? extends Component>> requiredComponents();
+    default List<Class<? extends Component>> readComponents() {
+        return List.of();
+    }
+    
+    default List<Class<? extends Component>> writeComponents() {
+        return List.of();
+    }
 }