+++ src/main/java/com/island/nature/service/AnimalHealthSystem.java
@@ -24,7 +24,12 @@ public class AnimalHealthSystem extends NatureEntitySystem {
     }
 
     @Override
-    public List<Class<? extends Component>> requiredComponents() {
+    public List<Class<? extends Component>> readComponents() {
+        return List.of();
+    }
+
+    @Override
+    public List<Class<? extends Component>> writeComponents() {
         return List.of(HealthComponent.class, AgeComponent.class, MetabolismComponent.class);
     }